run unittests in other projects (#502)
* coda-oss release 2022-05-03

* latest from coda-oss

* more updates from coda-oss

* fix build errors on Linux

* don't need VS UnitTest project

* latest from coda-oss

* latest from coda-oss

* trying to fix build error

* need to use "srand_called" to fix compiler warning

* latest from coda-oss

* latest from coda-oss

* coda-oss release 2022-06-29

* latest from coda-oss

* str::strip() is gone, use str::trim()

* copy export template from coda-oss

* coda-oss 2022-08-02

* latest from coda-oss

* latest from coda-oss

* latest from coda-oss

* coda-oss now has ToString() specialization

* finding the right paths can be a bit messy :-(

* utility routines for GIT root and NITRO root

* run NITRO unittests in SIX

* fix compiler warnngs

* get CMake unittests working w/o --install

* Squashed commit of the following:

commit 8582167dfbd4d7c5d2945a5def54802808bc8a78
Author: J. Daniel Smith <dan@jdanielsmith.me>
Date:   Tue Aug 30 09:53:56 2022 -0400

    coda-oss Release 2022-08-30

* remove C11 'support'

* remove more C11 "support"

* more C++14

* remove C11 "support"

* try harder to get the right NITF plugin path

* changes from coda

* remove unused code

* get CTests working on Linux

* looking for .git doesn't work on Windows

* try (again) to fix unittests on Windows

* make the C "test_tre_mods" unittest C++ so that the NITF_PLUGIN_PATH can be more easily set

* finish moving C test_tre_mods unittest

* set NITF_PLUGIN_PATH

* Squashed commit of the following:

commit f3e4ff4e4c35909d5f85bca2f473e9d8e379f51c
Author: Dan Smith <dan.smith@maxar.com>
Date:   Tue Sep 27 17:03:43 2022 -0400

    Update ReleaseNotes.md

commit f8d3fec58ca46c0ecd4b91356d0ee24f7b0f69d3
Author: Dan Smith <dan.smith@maxar.com>
Date:   Tue Sep 27 17:00:12 2022 -0400

    latest from coda-oss

commit 5111e5cf3a6a16fad4382a7899e0cd50237c77fe
Merge: 748cfae0e 0c6f81732
Author: Dan Smith <dan.smith@maxar.com>
Date:   Tue Sep 27 16:44:42 2022 -0400

    Merge branch 'main' into develop/update-coda-oss

commit 748cfae0e3b7d4950b49f11cd8c2e7112d62d4c9
Merge: 05916abaf 370221354
Author: Dan Smith <dan.smith@maxar.com>
Date:   Tue Sep 27 12:03:41 2022 -0400

    Merge branch 'main' into develop/update-coda-oss

commit 05916abaf5e0dfd54dc36e1955cc427ed2efb7d5
Merge: 7410082bb 8cda7dc49
Author: Dan Smith <dan.smith@maxar.com>
Date:   Wed Sep 21 14:48:31 2022 -0400

    Merge branch 'main' into develop/update-coda-oss

commit 7410082bbe658f4f286099c4934433cbcf583735
Author: Dan Smith <dan.smith@maxar.com>
Date:   Wed Sep 21 14:41:14 2022 -0400

    latest from coda-oss

commit fc2f0c68d473429e31076f022e2ccb89c927851a
Author: Dan Smith <dan.smith@maxar.com>
Date:   Wed Sep 21 14:36:41 2022 -0400

    Squashed commit of the following:

    commit e193cd1d384961f6368f3b42e4ccec1c2f6811fe
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Wed Sep 21 12:15:04 2022 -0400

        more C++14

    commit ac804d76aa5640a8824429658e21cca418278c5b
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Wed Sep 21 12:12:41 2022 -0400

        remove more C11 "support"

    commit 04586c762b4dd9857650f19482e41389aa64c53c
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Wed Sep 21 12:00:11 2022 -0400

        remove C11 'support'

    commit d2b4f35ef90c2979a52ff92db6b0d782efd73bfe
    Merge: 12e5bea99 72447c8d6
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Wed Sep 21 11:59:34 2022 -0400

        Merge branch 'main' into develop/unittests

    commit 12e5bea9901b44db2d6389d84a47bda1d564a908
    Merge: b05a15daa f0d801e95
    Author: J. Daniel Smith <dan@jdanielsmith.me>
    Date:   Tue Aug 30 10:35:56 2022 -0400

        Merge branch 'master' into develop/unittests

    commit b05a15daacafcec0ccd7f2b570a9d5b8968925c1
    Author: J. Daniel Smith <dan@jdanielsmith.me>
    Date:   Tue Aug 30 09:55:52 2022 -0400

        Squashed commit of the following:

        commit 8582167dfbd4d7c5d2945a5def54802808bc8a78
        Author: J. Daniel Smith <dan@jdanielsmith.me>
        Date:   Tue Aug 30 09:53:56 2022 -0400

            coda-oss Release 2022-08-30

    commit 115fe462fad36b91cc5095da15004dd251fb7243
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Mon Aug 29 12:50:24 2022 -0400

        get CMake unittests working w/o --install

    commit c130d8e8403219513d6eba3f1525efa3b4b9febd
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Mon Aug 29 12:09:16 2022 -0400

        fix compiler warnngs

    commit 3463780711257332c77747828825966efb81d01d
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Mon Aug 29 12:04:44 2022 -0400

        run NITRO unittests in SIX

    commit a888a71bfca5ba06895811e611f63316b19fb666
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Mon Aug 29 11:23:04 2022 -0400

        utility routines for GIT root and NITRO root

    commit 257047e615c0423e0f48fb1d7b0eb41cc3838b58
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Mon Aug 29 10:51:23 2022 -0400

        finding the right paths can be a bit messy :-(

    commit c45a6af62dad734d992a0d2cdd3d916739359c3e
    Author: Dan Smith <dan.smith@maxar.com>
    Date:   Mon Aug 29 10:38:59 2022 -0400

        coda-oss now has ToString() specialization

commit 00e83cbb7582dd9e01ba1ddef3b629ffd030aa06
Merge: 861155795 72447c8d6
Author: Dan Smith <dan.smith@maxar.com>
Date:   Wed Sep 21 14:36:31 2022 -0400

    Merge branch 'main' into develop/update-coda-oss

commit 861155795c52d103ab4c24c9f2e5b232c198328a
Author: J. Daniel Smith <dan@jdanielsmith.me>
Date:   Fri Sep 9 16:59:58 2022 -0400

    latest from coda-oss

commit 789f3bda0bacf527568a6ad8c327033a6f73156f
Author: Dan Smith <dan.smith@maxar.com>
Date:   Fri Sep 9 12:58:15 2022 -0400

    don't use details::toString_() as that breaks the buld

commit b5a511b006299f5cf4201166577d69a1a5c810f8
Author: Dan Smith <dan.smith@maxar.com>
Date:   Fri Sep 9 12:47:12 2022 -0400

    run CodeQL against "main", not "master"

commit 65459a15853c66e293a003e78b424693a34b98ea
Author: Dan Smith <dan.smith@maxar.com>
Date:   Fri Sep 9 12:25:40 2022 -0400

    YML for "main", not "master"

commit 52b25ae8f11c39b515c4d6dbee78f78dc48f47ac
Author: Dan Smith <dan.smith@maxar.com>
Date:   Fri Sep 9 12:12:16 2022 -0400

    remove str::toString<>() calls

commit 3b2c1b3a7b8a0d4ce27da9c598ab456ef81eb505
Author: Dan Smith <dan.smith@maxar.com>
Date:   Fri Sep 9 11:33:58 2022 -0400

    remove more C11 "support"

commit 95dd9c3a8d6ff12e9fb423be5215a346dec33dc6
Author: Dan Smith <dan.smith@maxar.com>
Date:   Fri Sep 9 11:24:11 2022 -0400

    don't try to use C11

commit ab24785116b6e483d3c64dd7a94e0b044186b105
Author: Dan Smith <dan.smith@maxar.com>
Date:   Fri Sep 9 10:31:57 2022 -0400

    fix new compiler warnings

commit 60c8245c97d70753fe55d98590275104d5057925
Author: Dan Smith <dan.smith@maxar.com>
Date:   Fri Sep 9 09:58:47 2022 -0400

    latest from coda-oss:main

commit e38b5a5c8e7048a87260377f8ef2b77e0e7ae356
Author: Dan Smith <dan.smith@maxar.com>
Date:   Fri Sep 9 09:54:18 2022 -0400

    merge main

commit a4efb2000933a1d9c62ac81b4f4aaa34ebdf9738
Author: J. Daniel Smith <dan@jdanielsmith.me>
Date:   Tue Aug 30 14:33:29 2022 -0400

    coda-oss 2022-08-30_cpp14

commit 5fe76dee778e427fb89c59144afd4b907c818557
Author: J. Daniel Smith <dan@jdanielsmith.me>
Date:   Tue Aug 30 14:30:31 2022 -0400

    slam in "main"

commit 8582167dfbd4d7c5d2945a5def54802808bc8a78
Author: J. Daniel Smith <dan@jdanielsmith.me>
Date:   Tue Aug 30 09:53:56 2022 -0400

    coda-oss Release 2022-08-30
