+++ Test/Test.vcxproj
@@ -68,7 +68,6 @@
       <EnforceTypeConversionRules>true</EnforceTypeConversionRules>
       <ConformanceMode>true</ConformanceMode>
       <DebugInformationFormat>ProgramDatabase</DebugInformationFormat>
-      <LanguageStandard_C>stdc11</LanguageStandard_C>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
@@ -90,7 +89,6 @@
       <ControlFlowGuard>Guard</ControlFlowGuard>
       <EnforceTypeConversionRules>true</EnforceTypeConversionRules>
       <ConformanceMode>true</ConformanceMode>
-      <LanguageStandard_C>stdc11</LanguageStandard_C>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
@@ -106,7 +104,6 @@
     <ClCompile Include="nitf_test_geo_utils.cpp" />
     <ClCompile Include="nitf_test_image_io.cpp" />
     <ClCompile Include="nitf_test_mem_source.cpp" />
-    <ClCompile Include="nitf_test_tre_mods.cpp" />
     <ClCompile Include="nitf_test_zero_field.cpp" />
     <ClCompile Include="nrt_test_buffer_adapter.cpp" />
     <ClCompile Include="nrt_test_core_values.cpp" />
