+++ modules/c++/nitf/CMakeLists.txt
@@ -105,6 +105,7 @@ coda_add_tests(
         test_image_segment_computer.cpp
         test_image_writer.cpp
         test_nitf_buffer_list.cpp
+        test_tre_mods.cpp
         test_tre_mods++.cpp
         test_tre_create++.cpp
         test_j2k_loading++.cpp
