+++ modules/c/nitf/CMakeLists.txt
@@ -80,7 +80,6 @@ coda_add_tests(
         test_image_io.c
         test_mem_source.c
         test_moveTREs.c
-        test_tre_mods.c
         test_zero_field.c
         )
